+++ tests/Pest.php
@@ -2,6 +2,9 @@
 
 declare(strict_types=1);
 
+use Emeq\HubSdk\Tests\RoutesEnabledTestCase;
 use Emeq\HubSdk\Tests\TestCase;
 
-uses(TestCase::class)->in(__DIR__);
+uses(TestCase::class)->in('Feature/HubClientTest.php');
+uses(TestCase::class)->in('Unit');
+uses(RoutesEnabledTestCase::class)->in('Feature/IntegrationRoutesTest.php');